Fixed problem with ResourceArc2
The problem with the  new custom connection (ResourceArc2) has been fixed, and now it is possible to distinguish two different  custom connections, one custom connection in red for all the elements except from Position which connection is black.

It was necessary to change types.js file and CustomRules.js  which contain functions to determine the custom connections of each custom element.

Finally, the function "drawConnection" of CustomRenderer has been changed to apply the two different colors depending on the type of the element.