Added first version of connector of can delegate
A new method has been added in CustomRules to render a complex connection with an image to show "delegate to"  connection. However it lacks of text.  Additionally it was necessary to modify "TimeSlot" to render the image of person in the new complex connection. Further modifications are required to create more custom connections.