Custom Renderer update
The custom render path variable has been updated in order to fix the connection problem within elements. Specifically, it has been added a better delimitation for each element, in form of path.

Additionally, a new function has been added to try to draw line with a crossed line, but it is not finished.