Fixed problem withj connections
Previosly, it was possible to connect two connections of an object to the same target, but now this issue has been fixed by checking the connections between the source of a connection and his target. Thus if the source has already been connected to the target, it is not possible to connect it again.