Fixed problem with  context pad icons and new conditions in ralph rules
The contextPad icons of the new hierarchy connectores were inverted in the case of delegation, now this issue has been fixed. Additionally, some conditions have been added for AND/Or and for history connectors to avoid problems. Specifically, the number of outcoming connections for AND/OR elements and dashed / solid connections for history connectors.